+++ pet-clinic-data/src/main/java/springframework/petclinic/model/Owner.java
@@ -1,0 +1,4 @@
+package springframework.petclinic.model;
+
+public class Owner extends Person {
+}
