+++ pet-clinic-data/src/main/java/springframework/petclinic/model/PetType.java
@@ -1,0 +1,14 @@
+package springframework.petclinic.model;
+
+public class PetType {
+
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
