+++ pet-clinic-web/src/main/java/springframework/petclinic/PetClinicApplication.java
@@ -1,0 +1,12 @@
+package springframework.petclinic;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PetClinicApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PetClinicApplication.class, args);
+    }
+}
