+++ pet-clinic-web/src/test/java/petclinic/PetClinicApplicationTests.java
@@ -1,0 +1,16 @@
+package petclinic;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class PetClinicApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
