--- src/main/java/np/info/ashok/petclinic/PetClinicApplication.java
@@ -1,12 +1,0 @@
-package np.info.ashok.petclinic;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PetClinicApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PetClinicApplication.class, args);
-    }
-}
