--- src/main/java/np/info/ashok/petclinic/model/Owner.java
@@ -1,4 +1,0 @@
-package np.info.ashok.petclinic.model;
-
-public class Owner extends Person {
-}
