--- src/main/java/np/info/ashok/petclinic/model/Person.java
@@ -1,23 +1,0 @@
-package np.info.ashok.petclinic.model;
-
-public class Person {
-
-    private String firstName;
-    private String lastName;
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-}
