--- src/main/java/np/info/ashok/petclinic/model/Pet.java
@@ -1,34 +1,0 @@
-package np.info.ashok.petclinic.model;
-
-import java.time.LocalDate;
-
-public class Pet extends Person {
-
-    private PetType petType;
-    private Owner owner;
-    private LocalDate birthDate;
-
-    public PetType getPetType() {
-        return petType;
-    }
-
-    public void setPetType(PetType petType) {
-        this.petType = petType;
-    }
-
-    public Owner getOwner() {
-        return owner;
-    }
-
-    public void setOwner(Owner owner) {
-        this.owner = owner;
-    }
-
-    public LocalDate getBirthDate() {
-        return birthDate;
-    }
-
-    public void setBirthDate(LocalDate birthDate) {
-        this.birthDate = birthDate;
-    }
-}
