--- src/main/java/np/info/ashok/petclinic/model/PetType.java
@@ -1,14 +1,0 @@
-package np.info.ashok.petclinic.model;
-
-public class PetType {
-
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
