--- src/main/java/np/info/ashok/petclinic/model/Vet.java
@@ -1,4 +1,0 @@
-package np.info.ashok.petclinic.model;
-
-public class Vet extends Person {
-}
